requirements.txt
absl-py @ file:///croot/absl-py_1714140470852/work
aiohappyeyeballs @ file:///croot/aiohappyeyeballs_1734469393482/work
aiohttp @ file:///croot/aiohttp_1734687138658/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
astunparse==1.6.3
async-timeout @ file:///croot/async-timeout_1732661977001/work
attrs @ file:///croot/attrs_1734533101012/work
blinker @ file:///croot/blinker_1737448726027/work
Brotli @ file:///croot/brotli-split_1736182456865/work
cachetools @ file:///croot/cachetools_1738224668060/work
certifi @ file:///croot/certifi_1745939216646/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///croot/click_1744271578095/work
cryptography @ file:///croot/cryptography_1694444244250/work
Flask @ file:///croot/flask_1737454299809/work
Flask-Cors @ file:///tmp/build/80754af9/flask_cors_1614875208036/work
flatbuffers @ file:///croot/python-flatbuffers_1722369094438/work
frozenlist @ file:///croot/frozenlist_1730902802621/work
gast @ file:///tmp/build/80754af9/gast_1637837472536/work
google-auth @ file:///croot/google-auth_1741275169457/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio @ file:///opt/conda/conda-bld/grpcio_1655801038495/work
h5py @ file:///croot/h5py_1730215935167/work
idna @ file:///croot/idna_1714398848350/work
itsdangerous @ file:///croot/itsdangerous_1716533337253/work
Jinja2 @ file:///croot/jinja2_1741710844255/work
keras @ file:///home/builder/stiwari/miniconda3/envs/tf_new_env/conda-bld/keras_1661556226316/work/keras-2.9.0-py2.py3-none-any.whl
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
Markdown @ file:///croot/markdown_1746114451035/work
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
multidict @ file:///croot/multidict_1730905498140/work
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1652801679809/work
oauthlib @ file:///croot/oauthlib_1679489621486/work
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging @ file:///croot/packaging_1734472117206/work
pillow @ file:///croot/pillow_1744613067434/work
propcache @ file:///croot/propcache_1744012718189/work
protobuf==3.20.3
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT @ file:///croot/pyjwt_1736178889029/work
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
requests @ file:///croot/requests_1730999120400/work
requests-oauthlib @ file:///croot/requests-oauthlib_1720615017878/work
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scipy @ file:///opt/conda/conda-bld/scipy_1661390393401/work
six @ file:///croot/six_1744271502820/work
tensorboard @ file:///home/builder/stiwari/miniconda3/envs/tf_new_env/conda-bld/tensorboard_1661447826088/work/tensorboard-2.9.0-py3-none-any.whl
tensorboard-data-server @ file:///croot/tensorboard-data-server_1670853588388/work/tensorboard_data_server-0.6.1-py3-none-manylinux2010_x86_64.whl
tensorboard-plugin-wit @ file:///home/builder/tkoch/workspace/tensorflow/tensorboard-plugin-wit_1658918494740/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow @ file:///home/builder/stiwari/miniconda3/envs/tfrebuild/conda-bld/tensorflow-base_1666072683393/work/tensorflow_pkg/tensorflow-2.9.1-cp310-cp310-linux_x86_64.whl
tensorflow-estimator @ file:///home/builder/stiwari/miniconda3/envs/tf_new_env/conda-bld/tensorflow-estimator_1661448080217/work/tensorflow_estimator-2.9.0-py2.py3-none-any.whl
termcolor @ file:///croot/termcolor_1668084651543/work
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
urllib3 @ file:///croot/urllib3_1737133630106/work
Werkzeug @ file:///croot/werkzeug_1737448706492/work
wrapt @ file:///croot/wrapt_1736540904746/work
yarl @ file:///croot/yarl_1732546845924/work
